#151 - Introducing IntegerNumberField
